requirements.txt
pandas==2.3.2
snowflake-snowpark-python
streamlit